backend/requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cmake==3.18.2.post1
defusedxml==0.6.0
Django==2.2.16
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
dlib==19.21.0
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mysqlclient==2.0.1
numpy==1.18.0
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
